# docs/openapi-summary.yaml
openapi: 3.0.3
info:
  title: SkillHub Go Edition API
  version: governance-complete-api-spec-v1
  description: |
    Summary OpenAPI for SkillHub. See docs/API_SPEC.md for detailed request/response schemas.
servers:
  - url: http://127.0.0.1:8848
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /v3/auth/setup:
    post:
      summary: First setup local admin
      responses:
        '200': { description: OK }
        '409': { description: Already initialized }
...
  /v3/admin/iam/local-users/{username}:
    put:
      summary: Update local user
      parameters:
        - in: path
          name: username
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
    delete:
      summary: Disable local user
      parameters:
        - in: path
          name: username
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
...
  /v3/client/ai/skills:
    get: { summary: Runtime download skill zip, responses: { '200': { description: ZIP }, '304': { description: Not modified } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key